pyproject.toml
[project]
name = "aatmiya-sabha-form"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "gspread>=6.2.1",
    "pandas>=2.3.1",
    "streamlit>=1.48.0",
]
